+++ ICD.Connect.Audio/VolumePoints/IVolumePointSettings.cs
@@ -45,8 +45,13 @@ namespace ICD.Connect.Audio.VolumePoints
 		long VolumeRampInitialInterval { get; set; }
 
 		/// <summary>
-		/// Gets/sets the context for the volume point.
+		/// Determines the contextual availability of this volume point.
 		/// </summary>
-		eVolumeType VolumeType { get; set; }
+		eVolumePointContext Context { get; set; }
+
+		/// <summary>
+		/// Determines what muting this volume point will do (mute audio output, mute microphones, etc).
+		/// </summary>
+		eMuteType MuteType { get; set; }
 	}
 }