+++ ICD.Connect.Audio/VolumePoints/eMuteType.cs
@@ -1,0 +1,20 @@
+﻿namespace ICD.Connect.Audio.VolumePoints
+{
+	public enum eMuteType
+	{
+		/// <summary>
+		/// Mutes the room audio output.
+		/// </summary>
+		RoomAudio,
+
+		/// <summary>
+		/// Mutes audio input at the DSP level.
+		/// </summary>
+		DspPrivacyMute,
+
+		/// <summary>
+		/// Mutes audio input at the microphone level.
+		/// </summary>
+		MicPrivacyMute
+	}
+}